gcp/src/requirements.txt
flask==3.0.3
functions-framework==3.7.0
google-cloud-aiplatform==1.53.0
google-cloud-documentai==2.28.0
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
retry2==0.9.5
redis==5.2.1
redisvl==0.4.1
numpy==2.2.4
openai==1.75.0